Finance.Belgium.StructuredCommunication

Description

Belgian companies often make use of structured communication with a checksum. This package aims to provide a toolkit to parse, render and manipulate StructuredCommunication.

Constructing StructuredCommunication

data StructuredCommunication

A data type that stores three numbers: one with three digits (000–999), four digits (0000–9999) and five digits (00001–99997). The data constructor itself is not accessible, since the StructuredCommunication could produce objects that are out of the given ranges, or where the checksum is not valid. The module thus aims to prevent parsing, changing, etc. StructuredCommunication objects into an invalid state.

Constructors

StructuredCommunication !Word16 !Word16 !Word32

structuredCommunication

Arguments

:: (Integral i, Integral j, Integral k)

=> i
The first number, should be between 000 and 999.

-> j
The second number, should be between 0000 and 9999.

-> k
The third number, should be between 00001 and 99997.

-> Maybe StructuredCommunication
The StructuredCommunication wrapped in a Just of the three numbers are in range, and the checksum matches, otherwise Nothing.

Construct a StructuredCommunication object for the given three integral values that form the three sequences of digits.

determining the checksum

checksum

Arguments

:: StructuredCommunication
The StructuredCommunication for which we determine the checkum.

-> Word32
The last two digits of the StructuredCommunication object. The checksum is not per se valid.

Determining the checksum-part for the given StructuredCommunication. This thus takes the last two digits, or the third number modulo one hundred.

determineChecksum

Arguments

:: StructuredCommunication
The StructuredCommunication to determine the checksum from.

-> Word32
The checksum determined by the first ten digits, not per se the real checksum of the StructuredCommunication.

Determine the checksum based on the first ten digits. If the StructuredCommunication is not valid, its checksum will not match the result of the determineChecksum.

validChecksum

Arguments

:: StructuredCommunication
The StructuredCommunication for which we check the checksum.

-> Bool
True if the checksum is valid; False otherwise.

Check if the checksum matches for the given StructuredCommunication.

fixChecksum

Arguments

:: StructuredCommunication
The given StructuredCommunication to fix.

-> StructuredCommunication
A variant of the given StructuredCommunication where only the last two digits are changed to have a valid checksum.

Convert the given StructuredCommunication to one where the checksum is valid. If the checksum was already valid, it returns an equivalent StructuredCommunication, this operation is thus idempotent.

Converting to text

communicationToString

Arguments

:: StructuredCommunication
The given StructuredCommunication to convert to a String.

-> String
The corresponding String, of the form +++000/0000/00097+++.

Convert the given StructuredCommunication to a String that looks like a structured communication, so +++000/0000/00097+++.

communicationToText

Arguments

:: StructuredCommunication
The given StructuredCommunication to convert to a Text.

-> Text
The corresponding Text, of the form +++000/0000/00097+++.

Convert the given StructuredCommunication to a Text that looks like a structured communication, so +++000/0000/00097+++.

Parsing from text

communicationParser

Arguments

:: Stream s m Char

=> ParsecT s u m StructuredCommunication
The ParsecT object that parses the structured communication of the form +++000/0000/00097+++.

A ParsecT that parses a string into a StructuredCommunication, the StructuredCommunication is checked for its validity (checksum). The parser does not (per se) ends with an eof.

communicationParser'

Arguments

:: Stream s m Char

=> ParsecT s u m StructuredCommunication
The ParsecT object that parses the structured communication of the form +++000/0000/00097+++.

A ParsecT that parses a string into a StructuredCommunication, the StructuredCommunication can be invalid. The parser also does not (per se) ends with an eof.

communicationEParser

Arguments

:: Stream s m Char

=> ParsecT s u m StructuredCommunication
The ParsecT object that parses the structured communication of the form +++000/0000/00097+++.

A ParsecT that parses a string into a StructuredCommunication, the StructuredCommunication is checked for its validity (checksum). The parser also checks that this is the end of the stream.

communicationEParser'

Arguments

:: Stream s m Char

=> ParsecT s u m StructuredCommunication
The ParsecT object that parses the structured communication of the form +++000/0000/00097+++.

A ParsecT that parses a string into a StructuredCommunication, the StructuredCommunication can be invalid. The parser also checks if this is the end of the stream.

parseCommunication

Arguments

:: Stream s Identity Char

=> s
The stream that is parsed into a StructuredCommunication

-> Either ParseError StructuredCommunication
The result of parsing, either a StructuredCommunication wrapped in a Right or a parsing error wrapped in a Left.

Parsing a stream into a StructuredCommunication that also validates the checksum of the communication. The stream does not per se needs to end with structured communcation.

parseCommunication'

Arguments

:: Stream s Identity Char

=> s
The stream that is parsed into a StructuredCommunication

-> Either ParseError StructuredCommunication
The result of parsing, either a StructuredCommunication wrapped in a Right or a parsing error wrapped in a Left.

Parsing a stream into a StructuredCommunication that does noet validate the checksum of the communication. The stream does not per se needs to end with structured communcation.

parseCommunicationE

Arguments

:: Stream s Identity Char

=> s
The stream that is parsed into a StructuredCommunication

-> Either ParseError StructuredCommunication
The result of parsing, either a StructuredCommunication wrapped in a Right or a parsing error wrapped in a Left.

Parsing a stream into a StructuredCommunication that also validates the checksum of the communication. After the structured communication, the stream needs to end.

parseCommunicationE'

Arguments

:: Stream s Identity Char

=> s
The stream that is parsed into a StructuredCommunication

-> Either ParseError StructuredCommunication
The result of parsing, either a StructuredCommunication wrapped in a Right or a parsing error wrapped in a Left.

Parsing a stream into a StructuredCommunication that does noet validate the checksum of the communication. After the structured communication, the stream needs to end.

Quasi quotation

beCommunication

Arguments

:: QuasiQuoter
A QuasiQuoter to parse to a StructuredCommunication.

A QuasiQuoter that can parse a string into an expression or pattern. It will thus convert +++000/000/00097+++ into a StructuredCommunication as expression or pattern.